LPS-57501 Use Comment API in Staging tests
